telegram-bot/requirements.txt
python-telegram-bot[ext]==21.10
supabase>=2.13.0
python-dotenv>=1.1.0
edge-tts>=7.0.0
aiofiles>=24.1.0
